+++ Data/Profunctor/Product/Default/Class.hs
@@ -36,16 +36,16 @@ type DefaultFields p a b = GDefCnstr p (Rep a) (Rep b)
 type DefaultFields' p a = DefaultFields p a a
 
 -- | @'DefaultPConstraints' p a@ expands to the minimal combination of
--- @'Profunctor' p@, @'ProductProfunctor' p@, @'SumProfunctor' p@ needed to implement
+-- @'Profunctor' p@, @'SemiproductProfunctor' p@, @'SumProfunctor' p@ needed to implement
 -- the instance @'Default' p a a@ for a 'Generic' datatype @a@.
 --
 -- > DefaultPConstraints p Foo =
--- >   ( ProductProfunctor p      -- because Foo has a constructor Bar with many fields
--- >   , SumProfunctor p          -- because Foo has multiple constructors
+-- >   ( SemiproductProfunctor p      -- because Foo has a constructor Bar with many fields
+-- >   , SumProfunctor p              -- because Foo has multiple constructors
 -- >   )
 --
 -- > DefaultConstraints p (a, b) =
--- >   ( ProductProfunctor p      -- (a, b) has a single constructor with two fields
+-- >   ( SemiproductProfunctor p      -- (a, b) has a single constructor with two fields
 -- >   )
 type DefaultPConstraints p a = GDefPCnstr p (Rep a)
 
@@ -74,7 +74,7 @@ class GDefault p f g where
   type GDefCnstr p f g :: Constraint
   gdef1 :: p (f a) (g a)
 
-instance ProductProfunctor p => GDefault p U1 U1 where
+instance SemiproductProfunctor p => GDefault p U1 U1 where
   type GDefCnstr p U1 U1 = ()
   gdef1 = dimap (const ()) (const U1) empty
 
@@ -86,7 +86,7 @@ instance (Profunctor p, Default p c c') => GDefault p (K1 i c) (K1 i c') where
   type GDefCnstr p (K1 i c) (K1 i c') = Default p c c'
   gdef1 = dimap unK1 K1 def
 
-instance (ProductProfunctor p, GDefault p f f', GDefault p g g') => GDefault p (f :*: g) (f' :*: g') where
+instance (SemiproductProfunctor p, GDefault p f f', GDefault p g g') => GDefault p (f :*: g) (f' :*: g') where
   type GDefCnstr p (f :*: g) (f' :*: g') = (GDefCnstr p f f', GDefCnstr p g g')
   gdef1 = dimap (\(x :*: y) -> (x, y)) (uncurry (:*:)) $ gdef1 ***! gdef1
 
@@ -102,10 +102,10 @@ instance (SumProfunctor p, GDefault p f f', GDefault p g g') => GDefault p (f :+
         R1 x -> Right x
 
 type family GDefPCnstr (p :: * -> * -> *) (f :: * -> *) :: Constraint
-type instance GDefPCnstr p U1 = ProductProfunctor p
+type instance GDefPCnstr p U1 = SemiproductProfunctor p
 type instance GDefPCnstr p (M1 i c f) = GDefPCnstr p f
 type instance GDefPCnstr p (K1 i c) = Profunctor p
-type instance GDefPCnstr p (f :*: g) = ProductProfunctor p
+type instance GDefPCnstr p (f :*: g) = SemiproductProfunctor p
 type instance GDefPCnstr p (f :+: g) = (SumProfunctor p, GDefPCnstr p f, GDefPCnstr p g)
 
 gdef :: (Profunctor p, Generic a, Generic b, GDefault p (Rep a) (Rep b)) => p a b